refactor: restructure dialog components and replace dialog-shell with DialogShell component
